extract frame definition from tests
